♻️ Remove backup.js and update next.config.js and .env files
Removed backup.js as it was not needed in the current deployment process. Updated next.config.js to enhance the export configuration and improve the website's performance. Updated .env files to secure the API keys by changing the tokens. These changes aim to optimize the performance and increase the security of the website.