🎨 Replace hard-coded URLs with environment variables
The code has been changed to use environment variables in multiple places instead of the previously hard-coded URLs. This is specifically for interactions with the WINDOWS_SERVER and DATOCMS_API. This will make the code more flexible and easier to manage, as the URLs can now be updated in a single location - the .env file - without having to make changes to the actual code.